Add vcat, hcat, and blkdiag with new signatures.
The new signatures allow the return type (which must be a
SparseMatrixCSC{Tv,Ti}) to be specified as a first argument and
gracefully handle the case of zero matrices to concatenate by returning
an empty matrix of the requested type. The methods without explicit
return type for hcat, vcat, and blkdiag now call the ones with explicit
return type after deducing the required types as before.